Gen complete + Save&Load

+++ PerlinNoise.h
@@ -23,7 +23,7 @@ class PerlinNoise {
             107, 49, 192, 214, 31, 181, 199, 106, 157, 184, 84, 204, 176, 115, 121, 50, 45, 127, 4, 150, 254,
             138, 236, 205, 93, 222, 114, 67, 29, 24, 72, 243, 141, 128, 195, 78, 66, 215, 61, 156, 180};
 public:
-    PerlinNoise(unsigned int seed) {
+    PerlinNoise(short seed) {
         p.resize(256);
 
         // Fill p with values from 0 to 255
